+++ lib/bin/require.rb
@@ -2,7 +2,7 @@
 
 class Require < BaseCommand
   def call(*args, env:, frames:)
-    path = args[0].pathmap('%d/%f')
+    path = Hal.realpath(args[0])   #.pathmap('%d/%f')
     if Hal.exist? path
       require path
       true
@@ -10,5 +10,9 @@ class Require < BaseCommand
       env[:err].puts "#{path}: File not found"
       false
     end
+  rescue SyntaxError => err
+    env[:err].puts "SyntaxError: #{err.message}"
+  rescue => err
+    env[:err].puts err.class.name, err.message
   end
 end